update: conv.ImgExt int -> img.Ext string; dirconv.Dirconv.Traverse.. -> dir.Traverse..

+++ kadai2/ebiiim/pkg/conv/conv.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/gopherdojo/dojo5/kadai2/ebiiim/pkg/img"
 	"github.com/pkg/errors"
 	"golang.org/x/image/bmp"
 	"golang.org/x/image/tiff"
@@ -18,11 +19,11 @@ type ImgConv struct {
 	// source path e.g. `path/to/A.jpg`
 	SrcPath string
 	// source image extension
-	SrcExt ImgExt
+	SrcExt img.Ext
 	// target path e.g. `path/to/A.png`
 	TgtPath string
 	// target image extension
-	TgtExt ImgExt
+	TgtExt img.Ext
 	// converter options (currently, this value is not interpreted by encodeImg())
 	Options map[string]interface{}
 }
@@ -52,49 +53,49 @@ func (ic *ImgConv) Convert() error {
 		return errors.Wrapf(err, "failed to create %s", ic.TgtPath)
 	}
 	err = encodeImg(tgtFile, ic.TgtExt, &srcImg, ic.Options)
+	cErr := tgtFile.Close() // close the file before it may be deleted.
+	if cErr != nil {
+		return errors.Wrapf(cErr, "failed to close %v", cErr)
+	}
 	if err != nil {
-		rErr := os.Remove(ic.TgtPath)
+		rErr := os.Remove(ic.TgtPath) // if failed to encode, delete the unnecessary zero-byte file
 		if rErr != nil {
 			err = errors.Wrapf(err, "failed to remove %v", rErr)
 		}
 		return errors.Wrapf(err, "failed to encode %s", ic.SrcPath)
 	}
-	err = tgtFile.Close()
-	if err != nil {
-		return errors.Wrapf(err, "failed to close %v", err)
-	}
 
 	return nil
 }
 
-func encodeImg(writer io.Writer, ext ImgExt, img *image.Image, options map[string]interface{}) (err error) {
+func encodeImg(writer io.Writer, ext img.Ext, image *image.Image, options map[string]interface{}) (err error) {
 	switch ext {
-	case ImgExtJPEG:
+	case img.JPEG:
 		// TODO: apply encoder options
-		err = jpeg.Encode(writer, *img, &jpeg.Options{})
-	case ImgExtPNG:
-		err = png.Encode(writer, *img)
-	case ImgExtTIFF:
+		err = jpeg.Encode(writer, *image, &jpeg.Options{})
+	case img.PNG:
+		err = png.Encode(writer, *image)
+	case img.TIFF:
 		// TODO: apply encoder options
-		err = tiff.Encode(writer, *img, &tiff.Options{})
-	case ImgExtBMP:
-		err = bmp.Encode(writer, *img)
+		err = tiff.Encode(writer, *image, &tiff.Options{})
+	case img.BMP:
+		err = bmp.Encode(writer, *image)
 	default:
 		err = fmt.Errorf("unsupported image extension %s", ext)
 	}
 	return
 }
 
-func decodeImg(reader io.Reader, ext ImgExt) (img image.Image, err error) {
+func decodeImg(reader io.Reader, ext img.Ext) (image image.Image, err error) {
 	switch ext {
-	case ImgExtJPEG:
-		img, err = jpeg.Decode(reader)
-	case ImgExtPNG:
-		img, err = png.Decode(reader)
-	case ImgExtTIFF:
-		img, err = tiff.Decode(reader)
-	case ImgExtBMP:
-		img, err = bmp.Decode(reader)
+	case img.JPEG:
+		image, err = jpeg.Decode(reader)
+	case img.PNG:
+		image, err = png.Decode(reader)
+	case img.TIFF:
+		image, err = tiff.Decode(reader)
+	case img.BMP:
+		image, err = bmp.Decode(reader)
 	default:
 		err = fmt.Errorf("unsupported image extension %s", ext)
 	}
